+++ src/mitim_modules/portals/utils/PORTALSinit.py
@@ -104,7 +104,7 @@ def initializeProblem(
     # Check if I will be able to calculate radiation
     speciesNotFound = []
     for i in range(len(profiles.Species)):
-        data_df = pd.read_csv(__mitimroot__ / "src" / "mitim_modules" / "powertorch" / "physics" / "radiation_chebyshev.csv")
+        data_df = pd.read_csv(__mitimroot__ / "src" / "mitim_modules" / "powertorch" / "physics_models" / "radiation_chebyshev.csv")
         if not (data_df['Ion'].str.lower()==profiles.Species[i]["N"].lower()).any():
             speciesNotFound.append(profiles.Species[i]["N"])
