+++ src/mitim_modules/powertorch/physics_models/parameterizers.py
@@ -1,0 +1,175 @@
+import copy
+import torch
+import numpy as np
+from mitim_modules.powertorch.utils import CALCtools
+from mitim_tools.misc_tools.LOGtools import printMsg as print
+from IPython import embed
+
+# <> Function to interpolate a curve <> 
+from mitim_tools.misc_tools.MATHtools import extrapolateCubicSpline as interpolation_function
+
+def piecewise_linear(
+    x_coord,
+    y_coord_raw,
+    x_coarse_tensor,
+    parameterize_in_aLx=True,
+    multiplier_quantity=1.0,
+    ):
+    """
+    Notes:
+        - x_coarse_tensor must be torch
+    """
+
+    # **********************************************************************************************************
+    # Define the integrator and derivator functions (based on whether I want to parameterize in aLx or in gradX)
+    # **********************************************************************************************************
+
+    if parameterize_in_aLx:
+        # 1/Lx = -1/X*dX/dr
+        integrator_function, derivator_function = (
+            CALCtools.integrateGradient,
+            CALCtools.produceGradient,
+        )
+    else:
+        # -dX/dr
+        integrator_function, derivator_function = (
+            CALCtools.integrateGradient_lin,
+            CALCtools.produceGradient_lin,
+        )
+
+    y_coord = torch.from_numpy(y_coord_raw).to(x_coarse_tensor) * multiplier_quantity
+
+    ygrad_coord = derivator_function( torch.from_numpy(x_coord).to(x_coarse_tensor), y_coord )
+
+    # **********************************************************************************************************
+    # Get control points
+    # **********************************************************************************************************
+
+    x_coarse = x_coarse_tensor[1:].cpu().numpy()
+
+    """
+    Define region to get control points from
+    ------------------------------------------------------------
+	Trick: Addition of extra point
+		This is important because if I don't, when I combine the trailing edge and the new
+		modified profile, there's going to be a discontinuity in the gradient.
+	"""
+    
+    ir_end = np.argmin(np.abs(x_coord - x_coarse[-1]))
+
+    if ir_end < len(x_coord) - 1:
+        ir = ir_end + 2  # To prevent that TGYRO does a 2nd order derivative
+        x_coarse = np.append(x_coarse, [x_coord[ir]])
+    else:
+        ir = ir_end
+
+	# Definition of trailing edge. Any point after, and including, the extra point
+    x_trail = torch.from_numpy(x_coord[ir:]).to(x_coarse_tensor)
+    y_trail = y_coord[ir:]
+    x_notrail = torch.from_numpy(x_coord[: ir + 1]).to(x_coarse_tensor)
+
+    # Produce control points, including a zero at the beginning
+    aLy_coarse = [[0.0, 0.0]]
+    for cont, i in enumerate(x_coarse):
+        yValue = ygrad_coord[np.argmin(np.abs(x_coord - i))]
+        aLy_coarse.append([i, yValue.cpu().item()])
+
+    aLy_coarse = torch.from_numpy(np.array(aLy_coarse)).to(ygrad_coord)
+
+    # Since the last one is an extra point very close, I'm making it the same
+    aLy_coarse[-1, 1] = aLy_coarse[-2, 1]
+
+    # Boundary condition at point moved by gridPointsAllowed
+    y_bc = torch.from_numpy(interpolation_function([x_coarse[-1]], x_coord, y_coord.cpu().numpy())).to(ygrad_coord)
+
+    # Boundary condition at point (ACTUAL THAT I WANT to keep fixed, i.e. rho=0.8)
+    y_bc_real = torch.from_numpy(interpolation_function([x_coarse[-2]], x_coord, y_coord.cpu().numpy())).to(ygrad_coord)
+
+    # **********************************************************************************************************
+    # Define profile_constructor functions
+    # **********************************************************************************************************
+
+    def profile_constructor_coarse(x, y, multiplier=multiplier_quantity):
+        """
+        Construct curve in a coarse grid
+        ----------------------------------------------------------------------------------------------------
+        This constructs a curve in any grid, with any batch given in y=y.
+        Useful for surrogate evaluations. Fast in a coarse grid. For HF evaluations,
+        I need to do in a finer grid so that it is consistent with TGYRO.
+        x, y must be (batch, radii),	y_bc must be (1)
+        """
+        return x, integrator_function(x, y, y_bc_real) / multiplier
+
+    def profile_constructor_middle(x, y, multiplier=multiplier_quantity):
+        """
+        Deparamterizes a finer profile based on the values in the coarse.
+        Reason why something like this is not used for the full profile is because derivative of this will not be as original,
+                which is needed to match TGYRO
+        """
+        yCPs = CALCtools.Interp1d()(aLy_coarse[:, 0][:-1].repeat((y.shape[0], 1)), y, x)
+        return x, integrator_function(x, yCPs, y_bc_real) / multiplier
+
+    def profile_constructor_fine(x, y, multiplier=multiplier_quantity):
+        """
+        Notes:
+            - x is a 1D array, but y can be a 2D array for a batch of individuals: (batch,x)
+            - I am assuming it is 1/LT for parameterization, but gives T
+        """
+
+        y = torch.atleast_2d(y)
+        x = x[0, :] if x.dim() == 2 else x
+
+        # Add the extra trick point
+        x = torch.cat((x, aLy_coarse[-1][0].repeat((1))))
+        y = torch.cat((y, aLy_coarse[-1][-1].repeat((y.shape[0], 1))), dim=1)
+
+        # Model curve (basically, what happens in between points)
+        yBS = CALCtools.Interp1d()(x.repeat(y.shape[0], 1), y, x_notrail.repeat(y.shape[0], 1))
+
+        """
+        ---------------------------------------------------------------------------------------------------------
+            Trick 1: smoothAroundCoarsing
+                TGYRO will use a 2nd order scheme to obtain gradients out of the profile, so a piecewise linear
+                will simply not give the right derivatives.
+                Here, this rough trick is to modify the points in gradient space around the coarse grid with the
+                same value of gradient, so in principle it doesn't matter the order of the derivative.
+        """
+        num_around = 1
+        for i in range(x.shape[0] - 2):
+            ir = torch.argmin(torch.abs(x[i + 1] - x_notrail))
+            for k in range(-num_around, num_around + 1, 1):
+                yBS[:, ir + k] = yBS[:, ir]
+        # --------------------------------------------------------------------------------------------------------
+
+        yBS = integrator_function(x_notrail.repeat(yBS.shape[0], 1), yBS.clone(), y_bc)
+
+        """
+        Trick 2: Correct y_bc
+            The y_bc for the profile integration started at gridPointsAllowed, but that's not the real
+            y_bc. I want the temperature fixed at my first point that I actually care for.
+            Here, I multiply the profile to get that.
+            Multiplication works because:
+                1/LT = 1/T * dT/dr
+                1/LT' = 1/(T*m) * d(T*m)/dr = 1/T * dT/dr = 1/LT
+            Same logarithmic gradient, but with the right boundary condition
+
+        """
+        ir = torch.argmin(torch.abs(x_notrail - x[-2]))
+        yBS = yBS * torch.transpose((y_bc_real / yBS[:, ir]).repeat(yBS.shape[1], 1), 0, 1)
+
+        # Add trailing edge
+        y_trailnew = copy.deepcopy(y_trail).repeat(yBS.shape[0], 1)
+
+        x_notrail_t = torch.cat((x_notrail[:-1], x_trail), dim=0)
+        yBS = torch.cat((yBS[:, :-1], y_trailnew), dim=1)
+
+        return x_notrail_t, yBS / multiplier
+
+    # **********************************************************************************************************
+
+    return (
+        aLy_coarse,
+        profile_constructor_fine,
+        profile_constructor_coarse,
+        profile_constructor_middle,
+    )
