+++ src/mitim_modules/powertorch/physics_models/transport_gacode.py
@@ -2,154 +2,18 @@ import copy
 import shutil
 import torch
 import numpy as np
-from mitim_tools.misc_tools import PLASMAtools, IOtools
-from mitim_tools.gacode_tools import TGYROtools, PROFILEStools
+from mitim_tools.misc_tools import IOtools
+from mitim_tools.gacode_tools import TGYROtools
 from mitim_modules.portals.utils import PORTALScgyro
+from mitim_modules.powertorch.utils import TRANSPORTtools
 from mitim_tools.misc_tools.LOGtools import printMsg as print
 from IPython import embed
-
-class power_transport:
-    '''
-    Default class for power transport models, change "evaluate" method to implement a new model and produce_profiles if the model requires written input.gacode written
-
-    Notes:
-        - After evaluation, the self.model_results attribute will contain the results of the model, which can be used for plotting and analysis
-        - model results can have .plot() method that can grab kwargs or be similar to TGYRO plot
-
-    '''
-    def __init__(self, powerstate, name = "test", folder = "~/scratch/", evaluation_number = 0):
-
-        self.name = name
-        self.folder = IOtools.expandPath(folder)
-        self.evaluation_number = evaluation_number
-        self.powerstate = powerstate
-
-        # Allowed fluxes in powerstate so far
-        self.quantities = ['Pe', 'Pi', 'Ce', 'CZ', 'Mt']
-
-        # Each flux has a turbulent and neoclassical component
-        self.variables = [f'{i}_tr_turb' for i in self.quantities] + [f'{i}_tr_neo' for i in self.quantities]
-
-        # Each flux component has a standard deviation
-        self.variables += [f'{i}_stds' for i in self.variables]
-
-        # There is also turbulent exchange
-        self.variables += ['PexchTurb', 'PexchTurb_stds']
-
-        # And total transport flux
-        self.variables += [f'{i}_tr' for i in self.quantities]
-
-        # Model results is None by default, but can be assigned in evaluate
-        self.model_results = None
-
-        # Assign zeros to transport ones if not evaluated
-        for i in self.variables:
-            self.powerstate.plasma[i] = self.powerstate.plasma["te"] * 0.0
-
-        # There is also target components
-        self.variables += [f'{i}' for i in self.quantities] + [f'{i}_stds' for i in self.quantities]
-
-        # ----------------------------------------------------------------------------------------
-        # labels for plotting
-        # ----------------------------------------------------------------------------------------
-
-        self.powerstate.labelsFluxes = {
-            "te": "$Q_e$ ($MW/m^2$)",
-            "ti": "$Q_i$ ($MW/m^2$)",
-            "ne": (
-                "$Q_{conv}$ ($MW/m^2$)"
-                if self.powerstate.TransportOptions["ModelOptions"].get("useConvectiveFluxes", True)
-                else "$\\Gamma_e$ ($10^{20}/s/m^2$)"
-            ),
-            "nZ": (
-                "$Q_{conv}$ $\\cdot f_{Z,0}$ ($MW/m^2$)"
-                if self.powerstate.TransportOptions["ModelOptions"].get("useConvectiveFluxes", True)
-                else "$\\Gamma_Z$ $\\cdot f_{Z,0}$ ($10^{20}/s/m^2$)"
-            ),
-            "w0": "$M_T$ ($J/m^2$)",
-        }
-
-    def produce_profiles(self):
-        # Only add self._produce_profiles() if it's needed (e.g. full TGLF), otherwise this is somewhat expensive (e.g. for flux matching)
-        pass
-
-    def _produce_profiles(self,deriveQuantities=True):
-
-        self.applyCorrections = self.powerstate.TransportOptions["ModelOptions"].get("MODELparameters", {}).get("applyCorrections", {})
-
-        # Write this updated profiles class (with parameterized profiles and target powers)
-        self.file_profs = self.folder / "input.gacode"
-
-        powerstate_detached = self.powerstate.copy_state()
-
-        self.powerstate.profiles = powerstate_detached.from_powerstate(
-            write_input_gacode=self.file_profs,
-            postprocess_input_gacode=self.applyCorrections,
-            rederive_profiles = deriveQuantities,        # Derive quantities so that it's ready for analysis and plotting later
-            insert_highres_powers = deriveQuantities,    # Insert powers so that Q, Pfus and all that it's consistent when read later
-        )
-
-        self.profiles_transport = copy.deepcopy(self.powerstate.profiles)
-
-        self._modify_profiles()
-
-    def _modify_profiles(self):
-        '''
-        Modify the profiles (e.g. lumping) before running the transport model 
-        '''
-
-        # After producing the profiles, copy for future modifications
-        self.file_profs_unmod = self.file_profs.parent / f"{self.file_profs.name}_unmodified"
-        shutil.copy2(self.file_profs, self.file_profs_unmod)
-
-        profiles_postprocessing_fun = self.powerstate.TransportOptions["ModelOptions"].get("profiles_postprocessing_fun", None)
-
-        if profiles_postprocessing_fun is not None:
-            print(f"\t- Modifying input.gacode to run transport calculations based on {profiles_postprocessing_fun}",typeMsg="i")
-            self.profiles_transport = profiles_postprocessing_fun(self.file_profs)
-
-        # Position of impurity ion may have changed
-        p_old = PROFILEStools.PROFILES_GACODE(self.file_profs_unmod)
-        p_new = PROFILEStools.PROFILES_GACODE(self.file_profs)
-
-        impurity_of_interest = p_old.Species[self.powerstate.impurityPosition]
-
-        try:
-            impurityPosition_new = p_new.Species.index(impurity_of_interest)
-
-        except ValueError:
-            print(f"\t- Impurity {impurity_of_interest} not found in new profiles, keeping position {self.powerstate.impurityPosition}",typeMsg="w")
-            impurityPosition_new = self.powerstate.impurityPosition
-
-        if impurityPosition_new != self.powerstate.impurityPosition:
-            print(f"\t- Impurity position has changed from {self.powerstate.impurityPosition} to {impurityPosition_new}",typeMsg="w")
-            self.powerstate.impurityPosition_transport = p_new.Species.index(impurity_of_interest)
-
-    # ----------------------------------------------------------------------------------------------------
-    # EVALUATE (custom part)
-    # ----------------------------------------------------------------------------------------------------
-    def evaluate(self):
-        '''
-        This needs to populate the following in self.powerstate.plasma
-            - Pe, Pe_tr, Pe_tr_turb, Pe_tr_neo -> MW/m^2
-            - Pi, Pi_tr, Pi_tr_turb, Pi_tr_neo -> MW/m^2
-            - Ce, Ce_tr, Ce_tr_turb, Ce_tr_neo -> MW/m^2
-                * Ce_raw, Ce_raw_tr, Ce_raw_tr_turb, Ce_raw_tr_neo -> 10^20/s/m^2
-            - CZ, CZ_tr, CZ_tr_turb, CZ_tr_neo -> MW/m^2 (but modified as needed, for example dividing by fZ0)
-                * CZ_raw, CZ_raw_tr, CZ_raw_tr_turb, CZ_raw_tr_neo -> 10^20/s/m^2  (NOT modified)
-            - Mt, Mt_tr, Mt_tr_turb, Mt_tr_neo -> J/m^2
-            - PexchTurb -> MW/m^3
-        and their respective standard deviations
-        '''
-
-        print(">> No transport fluxes to evaluate", typeMsg="w")
-        pass
 
 # ----------------------------------------------------------------------------------------------------
 # FULL TGYRO
 # ----------------------------------------------------------------------------------------------------
 
-class tgyro_model(power_transport):
+class tgyro_model(TRANSPORTtools.power_transport):
     def __init__(self, powerstate, **kwargs):
         super().__init__(powerstate, **kwargs)
 
@@ -514,90 +378,6 @@ def tglf_scan_trick(
     return Qe_point, Qi_point, Ge_point, GZ_point, Mt_point, Pexch_point, Qe_std, Qi_std, Ge_std, GZ_std, Mt_std, Pexch_std
 
 
-# ------------------------------------------------------------------
-# SIMPLE Diffusion (#TODO: implement with particle flux and the raw)
-# ------------------------------------------------------------------
-
-class diffusion_model(power_transport):
-    def __init__(self, powerstate, **kwargs):
-        super().__init__(powerstate, **kwargs)
-
-        # Ensure that the provided diffusivities include the zero location
-        self.chi_e = self.powerstate.TransportOptions["ModelOptions"]["chi_e"]
-        self.chi_i = self.powerstate.TransportOptions["ModelOptions"]["chi_i"]
-
-        if self.chi_e.shape[0] < self.powerstate.plasma['rho'].shape[-1]:
-            self.chi_e = torch.cat((torch.zeros(1), self.chi_e))
-
-        if self.chi_i.shape[0] < self.powerstate.plasma['rho'].shape[-1]:
-            self.chi_i = torch.cat((torch.zeros(1), self.chi_i))
-
-    def produce_profiles(self):
-        pass
-
-    def evaluate(self):
-
-        # Make sure the chis are applied to all the points in the batch
-        Pe_tr = PLASMAtools.conduction(
-            self.powerstate.plasma["ne"],
-            self.powerstate.plasma["te"],
-            self.chi_e.repeat(self.powerstate.plasma['rho'].shape[0],1),
-            self.powerstate.plasma["aLte"],
-            self.powerstate.plasma["a"].unsqueeze(-1),
-        )
-        Pi_tr = PLASMAtools.conduction(
-            self.powerstate.plasma["ni"].sum(axis=-1),
-            self.powerstate.plasma["ti"],
-            self.chi_i.repeat(self.powerstate.plasma['rho'].shape[0],1),
-            self.powerstate.plasma["aLti"],
-            self.powerstate.plasma["a"].unsqueeze(-1),
-        )
-
-        self.powerstate.plasma["Pe_tr_turb"] = Pe_tr * 2 / 3
-        self.powerstate.plasma["Pi_tr_turb"] = Pi_tr * 2 / 3
-
-        self.powerstate.plasma["Pe_tr_neo"] = Pe_tr * 1 / 3
-        self.powerstate.plasma["Pi_tr_neo"] = Pi_tr * 1 / 3
-
-        self.powerstate.plasma["Pe_tr"] = self.powerstate.plasma["Pe_tr_turb"] + self.powerstate.plasma["Pe_tr_neo"]
-        self.powerstate.plasma["Pi_tr"] = self.powerstate.plasma["Pi_tr_turb"] + self.powerstate.plasma["Pi_tr_neo"]
-
-# ------------------------------------------------------------------
-# SURROGATE
-# ------------------------------------------------------------------
-
-class surrogate_model(power_transport):
-    def __init__(self, powerstate, **kwargs):
-        super().__init__(powerstate, **kwargs)
-
-    def produce_profiles(self):
-        pass
-
-    def evaluate(self):
-
-        """
-        flux_fun as given in ModelOptions must produce Q and Qtargets in order of te,ti,ne
-        """
-
-        X = torch.Tensor()
-        for prof in self.powerstate.ProfilesPredicted:
-            X = torch.cat((X,self.powerstate.plasma['aL'+prof][:,1:]),axis=1)
-
-        _, Q, _, _ = self.powerstate.TransportOptions["ModelOptions"]["flux_fun"](X)
-
-        numeach = self.powerstate.plasma["rho"].shape[1] - 1
-
-        quantities = {
-            "te": "Pe",
-            "ti": "Pi",
-            "ne": "Ce",
-            "nZ": "CZ",
-            "w0": "Mt",
-        }
-
-        for c, i in enumerate(self.powerstate.ProfilesPredicted):
-            self.powerstate.plasma[f"{quantities[i]}_tr"] = torch.cat((torch.tensor([[0.0]]),Q[:, numeach * c : numeach * (c + 1)]),dim=1)
-
 # **************************************************************************************************
 # Functions
 # **************************************************************************************************
@@ -879,3 +659,4 @@ def dummyCDF(GeneralFolder, FolderEvaluation):
     cdf = FolderEvaluation / f"{subname}_ev{name}.CDF"
 
     return cdf
+
