+++ src/mitim_modules/powertorch/utils/TRANSFORMtools.py
@@ -3,10 +3,9 @@ import torch
 from pathlib import Path
 import numpy as np
 import pandas as pd
-from mitim_modules.powertorch.physics import CALCtools
 from mitim_tools.misc_tools import LOGtools, IOtools
 from mitim_tools.gacode_tools import PROFILEStools
-from mitim_modules.powertorch.physics import TARGETStools
+from mitim_modules.powertorch.physics_models import targets_analytic, parameterizers
 from mitim_tools.misc_tools.LOGtools import printMsg as print
 from mitim_tools import __mitimroot__
 from IPython import embed
@@ -14,7 +13,7 @@ from IPython import embed
 # <> Function to interpolate a curve <> 
 from mitim_tools.misc_tools.MATHtools import extrapolateCubicSpline as interpolation_function
 
-def gacode_to_powerstate(self, increase_profile_resol=False):
+def gacode_to_powerstate(self, rho_vec=None):
     """
     This function converts from the fine input.gacode grid to a powertorch object and grid.
     Notes:
@@ -31,12 +30,9 @@ def gacode_to_powerstate(self, increase_profile_resol=False):
 
     print("\t- Producing powerstate object from input.gacode")
 
-    input_gacode    = self.profiles
-    rho_vec         = self.plasma["rho"]
-
-    # Resolution of input.gacode
-    if increase_profile_resol:
-        improve_resolution_profiles(input_gacode, rho_vec)
+    input_gacode = self.profiles
+    if rho_vec is None:
+        rho_vec = self.plasma["rho"]
 
     # *********************************************************************************************
     # Radial grid
@@ -183,16 +179,16 @@ def gacode_to_powerstate(self, increase_profile_resol=False):
     for i in range(input_gacode.profiles['ni(10^19/m^3)'].shape[1]):
         cases_to_parameterize.append([f"ni{i}", "ni(10^19/m^3)", i, 1.0, True])
 
-    self.deparametrizers_fine, self.deparametrizers_coarse, self.deparametrizers_coarse_middle = {}, {}, {}
+    self.profile_constructors_fine, self.profile_constructors_coarse, self.profile_constructors_coarse_middle = {}, {}, {}
     for key in cases_to_parameterize:
         quant = input_gacode.profiles[key[1]] if key[2] is None else input_gacode.profiles[key[1]][:, key[2]]
 
         (
             aLy_coarse,
-            self.deparametrizers_fine[key[0]],
-            self.deparametrizers_coarse[key[0]],
-            self.deparametrizers_coarse_middle[key[0]],
-        ) = parameterize_profile(
+            self.profile_constructors_fine[key[0]],
+            self.profile_constructors_coarse[key[0]],
+            self.profile_constructors_coarse_middle[key[0]],
+        ) = parameterizers.piecewise_linear(
             input_gacode.derived["roa"],
             quant,
             self.plasma["roa"],
@@ -209,44 +205,44 @@ def gacode_to_powerstate(self, increase_profile_resol=False):
                 print(f"\t- All values of {key[0]} detected to be zero, to avoid NaNs, inserting {addT} at the edge",typeMsg="w")
                 self.plasma[f"aL{key[0]}"][..., -1] += addT
 
-    def to_gacode(
+def to_gacode(
+    self,
+    write_input_gacode=None,
+    position_in_powerstate_batch=0,
+    postprocess_input_gacode={},
+    insert_highres_powers=False,
+    rederive_profiles=True,
+):
+    '''
+    Notes:
+        - insert_highres_powers: whether to insert high resolution powers (will calculate them with powerstate targets object, not other custom ones)
+    '''
+    print(">> Inserting powerstate into input.gacode")
+
+    profiles = powerstate_to_gacode(
         self,
-        write_input_gacode=None,
-        position_in_powerstate_batch=0,
-        postprocess_input_gacode={},
-        insert_highres_powers=False,
-        rederive_profiles=True,
-    ):
-        '''
-        Notes:
-            - insert_highres_powers: whether to insert high resolution powers (will calculate them with powerstate targets object, not other custom ones)
-        '''
-        print(">> Inserting powerstate into input.gacode")
+        position_in_powerstate_batch=position_in_powerstate_batch,
+        postprocess_input_gacode=postprocess_input_gacode,
+        insert_highres_powers=insert_highres_powers,
+        rederive=rederive_profiles,
+    )
 
-        profiles = powerstate_to_gacode(
-            self,
-            position_in_powerstate_batch=position_in_powerstate_batch,
-            postprocess_input_gacode=postprocess_input_gacode,
-            insert_highres_powers=insert_highres_powers,
-            rederive=rederive_profiles,
-        )
+    # Write input.gacode
+    if write_input_gacode is not None:
+        write_input_gacode = Path(write_input_gacode)
+        print(f"\t- Writing input.gacode file: {IOtools.clipstr(write_input_gacode)}")
+        write_input_gacode.parent.mkdir(parents=True, exist_ok=True)
+        profiles.writeCurrentStatus(file=write_input_gacode)
 
-        # Write input.gacode
-        if write_input_gacode is not None:
-            write_input_gacode = Path(write_input_gacode)
-            print(f"\t- Writing input.gacode file: {IOtools.clipstr(write_input_gacode)}")
-            write_input_gacode.parent.mkdir(parents=True, exist_ok=True)
-            profiles.writeCurrentStatus(file=write_input_gacode)
+    # If corrections modify the ions set... it's better to re-read, otherwise powerstate will be confused
+    if rederive_profiles:
+        defineIons(self, profiles, self.plasma["rho"][position_in_powerstate_batch, :], self.dfT)
+        # Repeat, that's how it's done earlier
+        self._repeat_tensors(batch_size=self.plasma["rho"].shape[0],
+            specific_keys=["ni","ions_set_mi","ions_set_Zi","ions_set_Dion","ions_set_Tion","ions_set_c_rad"],
+            positionToUnrepeat=None)
 
-        # If corrections modify the ions set... it's better to re-read, otherwise powerstate will be confused
-        if rederive_profiles:
-            defineIons(self, profiles, self.plasma["rho"][position_in_powerstate_batch, :], self.dfT)
-            # Repeat, that's how it's done earlier
-            self._repeat_tensors(batch_size=self.plasma["rho"].shape[0],
-                specific_keys=["ni","ions_set_mi","ions_set_Zi","ions_set_Dion","ions_set_Tion","ions_set_c_rad"],
-                positionToUnrepeat=None)
-
-        return profiles
+    return profiles
 
 def powerstate_to_gacode(
     self,
@@ -294,7 +290,7 @@ def powerstate_to_gacode(
             # *********************************************************************************************
             # From a/Lx to x via fine profile_constructor
             # *********************************************************************************************
-            x, y = self.deparametrizers_fine[key[0]](
+            x, y = self.profile_constructors_fine[key[0]](
                 self.plasma["roa"][position_in_powerstate_batch, :],
                 self.plasma[f"aL{key[0]}"][position_in_powerstate_batch, :],
             )
@@ -375,7 +371,7 @@ def powerstate_to_gacode_powers(self, profiles, position_in_powerstate_batch=0):
             profiles,
             EvolutionOptions={"rhoPredicted": rhoy},
             TargetOptions={
-                "targets_evaluator": TARGETStools.analytical_model,
+                "targets_evaluator": targets_analytic.analytical_model,
                 "ModelOptions": {
                     "TypeTarget": self.TargetOptions["ModelOptions"]["TypeTarget"], # Important to keep the same as in the original
                     "TargetCalc": "powerstate",
@@ -433,7 +429,7 @@ def defineIons(self, input_gacode, rho_vec, dfT):
             Zi.append(input_gacode.profiles["z"][i])
 
             # Grab chebyshev coefficients from file
-            data_df = pd.read_csv(__mitimroot__ / "src" / "mitim_modules" / "powertorch" / "physics" / "radiation_chebyshev.csv")
+            data_df = pd.read_csv(__mitimroot__ / "src" / "mitim_modules" / "powertorch" / "physics_models" / "radiation_chebyshev.csv")
             try:
                 c = data_df[data_df['Ion'].str.lower()==input_gacode.profiles["name"][i].lower()].to_numpy()[0,2:].astype(float)
             except IndexError:
@@ -457,172 +453,6 @@ def defineIons(self, input_gacode, rho_vec, dfT):
     self.plasma["ions_set_Dion"] = Dion
     self.plasma["ions_set_Tion"] = Tion
     self.plasma["ions_set_c_rad"] = c_rad
-
-def parameterize_profile(
-    x_coord,
-    y_coord_raw,
-    x_coarse_tensor,
-    parameterize_in_aLx=True,
-    multiplier_quantity=1.0,
-    ):
-    """
-    Notes:
-        - x_coarse_tensor must be torch
-    """
-
-    # **********************************************************************************************************
-    # Define the integrator and derivator functions (based on whether I want to parameterize in aLx or in gradX)
-    # **********************************************************************************************************
-
-    if parameterize_in_aLx:
-        # 1/Lx = -1/X*dX/dr
-        integrator_function, derivator_function = (
-            CALCtools.integrateGradient,
-            CALCtools.produceGradient,
-        )
-    else:
-        # -dX/dr
-        integrator_function, derivator_function = (
-            CALCtools.integrateGradient_lin,
-            CALCtools.produceGradient_lin,
-        )
-
-    y_coord = torch.from_numpy(y_coord_raw).to(x_coarse_tensor) * multiplier_quantity
-
-    ygrad_coord = derivator_function( torch.from_numpy(x_coord).to(x_coarse_tensor), y_coord )
-
-    # **********************************************************************************************************
-    # Get control points
-    # **********************************************************************************************************
-
-    x_coarse = x_coarse_tensor[1:].cpu().numpy()
-
-    """
-    Define region to get control points from
-    ------------------------------------------------------------
-	Trick: Addition of extra point
-		This is important because if I don't, when I combine the trailing edge and the new
-		modified profile, there's going to be a discontinuity in the gradient.
-	"""
-    
-    ir_end = np.argmin(np.abs(x_coord - x_coarse[-1]))
-
-    if ir_end < len(x_coord) - 1:
-        ir = ir_end + 2  # To prevent that TGYRO does a 2nd order derivative
-        x_coarse = np.append(x_coarse, [x_coord[ir]])
-    else:
-        ir = ir_end
-
-	# Definition of trailing edge. Any point after, and including, the extra point
-    x_trail = torch.from_numpy(x_coord[ir:]).to(x_coarse_tensor)
-    y_trail = y_coord[ir:]
-    x_notrail = torch.from_numpy(x_coord[: ir + 1]).to(x_coarse_tensor)
-
-    # Produce control points, including a zero at the beginning
-    aLy_coarse = [[0.0, 0.0]]
-    for cont, i in enumerate(x_coarse):
-        yValue = ygrad_coord[np.argmin(np.abs(x_coord - i))]
-        aLy_coarse.append([i, yValue.cpu().item()])
-
-    aLy_coarse = torch.from_numpy(np.array(aLy_coarse)).to(ygrad_coord)
-
-    # Since the last one is an extra point very close, I'm making it the same
-    aLy_coarse[-1, 1] = aLy_coarse[-2, 1]
-
-    # Boundary condition at point moved by gridPointsAllowed
-    y_bc = torch.from_numpy(interpolation_function([x_coarse[-1]], x_coord, y_coord.cpu().numpy())).to(ygrad_coord)
-
-    # Boundary condition at point (ACTUAL THAT I WANT to keep fixed, i.e. rho=0.8)
-    y_bc_real = torch.from_numpy(interpolation_function([x_coarse[-2]], x_coord, y_coord.cpu().numpy())).to(ygrad_coord)
-
-    # **********************************************************************************************************
-    # Define profile_constructor functions
-    # **********************************************************************************************************
-
-    def profile_constructor_coarse(x, y, multiplier=multiplier_quantity):
-        """
-        Construct curve in a coarse grid
-        ----------------------------------------------------------------------------------------------------
-        This constructs a curve in any grid, with any batch given in y=y.
-        Useful for surrogate evaluations. Fast in a coarse grid. For HF evaluations,
-        I need to do in a finer grid so that it is consistent with TGYRO.
-        x, y must be (batch, radii),	y_bc must be (1)
-        """
-        return x, integrator_function(x, y, y_bc_real) / multiplier
-
-    def profile_constructor_middle(x, y, multiplier=multiplier_quantity):
-        """
-        Deparamterizes a finer profile based on the values in the coarse.
-        Reason why something like this is not used for the full profile is because derivative of this will not be as original,
-                which is needed to match TGYRO
-        """
-        yCPs = CALCtools.Interp1d()(aLy_coarse[:, 0][:-1].repeat((y.shape[0], 1)), y, x)
-        return x, integrator_function(x, yCPs, y_bc_real) / multiplier
-
-    def profile_constructor_fine(x, y, multiplier=multiplier_quantity):
-        """
-        Notes:
-            - x is a 1D array, but y can be a 2D array for a batch of individuals: (batch,x)
-            - I am assuming it is 1/LT for parameterization, but gives T
-        """
-
-        y = torch.atleast_2d(y)
-        x = x[0, :] if x.dim() == 2 else x
-
-        # Add the extra trick point
-        x = torch.cat((x, aLy_coarse[-1][0].repeat((1))))
-        y = torch.cat((y, aLy_coarse[-1][-1].repeat((y.shape[0], 1))), dim=1)
-
-        # Model curve (basically, what happens in between points)
-        yBS = CALCtools.Interp1d()(x.repeat(y.shape[0], 1), y, x_notrail.repeat(y.shape[0], 1))
-
-        """
-        ---------------------------------------------------------------------------------------------------------
-            Trick 1: smoothAroundCoarsing
-                TGYRO will use a 2nd order scheme to obtain gradients out of the profile, so a piecewise linear
-                will simply not give the right derivatives.
-                Here, this rough trick is to modify the points in gradient space around the coarse grid with the
-                same value of gradient, so in principle it doesn't matter the order of the derivative.
-        """
-        num_around = 1
-        for i in range(x.shape[0] - 2):
-            ir = torch.argmin(torch.abs(x[i + 1] - x_notrail))
-            for k in range(-num_around, num_around + 1, 1):
-                yBS[:, ir + k] = yBS[:, ir]
-        # --------------------------------------------------------------------------------------------------------
-
-        yBS = integrator_function(x_notrail.repeat(yBS.shape[0], 1), yBS.clone(), y_bc)
-
-        """
-        Trick 2: Correct y_bc
-            The y_bc for the profile integration started at gridPointsAllowed, but that's not the real
-            y_bc. I want the temperature fixed at my first point that I actually care for.
-            Here, I multiply the profile to get that.
-            Multiplication works because:
-                1/LT = 1/T * dT/dr
-                1/LT' = 1/(T*m) * d(T*m)/dr = 1/T * dT/dr = 1/LT
-            Same logarithmic gradient, but with the right boundary condition
-
-        """
-        ir = torch.argmin(torch.abs(x_notrail - x[-2]))
-        yBS = yBS * torch.transpose((y_bc_real / yBS[:, ir]).repeat(yBS.shape[1], 1), 0, 1)
-
-        # Add trailing edge
-        y_trailnew = copy.deepcopy(y_trail).repeat(yBS.shape[0], 1)
-
-        x_notrail_t = torch.cat((x_notrail[:-1], x_trail), dim=0)
-        yBS = torch.cat((yBS[:, :-1], y_trailnew), dim=1)
-
-        return x_notrail_t, yBS / multiplier
-
-    # **********************************************************************************************************
-
-    return (
-        aLy_coarse,
-        profile_constructor_fine,
-        profile_constructor_coarse,
-        profile_constructor_middle,
-    )
 
 def improve_resolution_profiles(profiles, rhoMODEL):
     """
@@ -673,7 +503,6 @@ def improve_resolution_profiles(profiles, rhoMODEL):
     # Change resolution
     # ----------------------------------------------------------------------------------
     profiles.changeResolution(rho_new=rho_new)
-
 
 def debug_transformation(p, p_new, s):
