+++ src/mitim_tools/misc_tools/PLASMAtools.py
@@ -8,7 +8,7 @@ import numpy as np
 import matplotlib.pyplot as plt
 from IPython import embed
 from mitim_tools.misc_tools import MATHtools
-from mitim_modules.powertorch.physics import CALCtools
+from mitim_modules.powertorch.utils import CALCtools
 from mitim_tools.popcon_tools import FunctionalForms
 from mitim_tools.misc_tools.LOGtools import printMsg as print
